# contracts/openapi/latest/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "kind": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "UIModelDataResponse": {
        "additionalProperties": true,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "description": "UI model item",
              "type": "object"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OpenAPI document for the runtime generic platform APIs. Registered modules become available through the generic model, document, UI contract, search, and offline endpoints. Custom module-specific controller namespaces are not auto-generated.",
    "title": "Orbyte Platform API",
    "version": "0.1.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ui/data/models": {
      "get": {
        "description": "Returns model data for the generic user workspace.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UIModelDataResponse"
                }
              }
            },
            "description": "UI model data"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Get model list for UI",
        "tags": [
          "UI Contracts"
        ],
        "x-orbyte-support-level": "public-headless"
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}